requirements.txt
Flask == 0.12.2
gevent
keras
numpy
h5py
pillow
scikit-learn
tensorflow == 1.6.0rc1